Changed to microsecond accuracy in the system.
Timing was moved off from the preemption module to its own, now in lib/timer.* - I also #if'd out the whole of the preemption module so that it is only compiled in if preemption is enabled (it's still broken!). Also modified a couple of defaults - the standard timeslice - and the granularity of the fair scheduler. Times are now 64-bits, even on 32-bit targets. Dead sleeping actually required no changes - it was broken as-is and interpreted the millisecond-accurate dead sleep time as microseconds.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>