Refactored schedulers.
Removed the RB, multilevel and random schedulers which were all either (a) crappy or (b) broken - or (c) both of the above. Rewrote sched_roundrobin from scratch using fifos. Refactored rescheduling/unscheduling into separate functions. Renamed the scheduler insert operation to 'schedule'.

Instead of having a rough API in scheduler.h implemented by whatever chosen scheduler, instead made struct scheduler contain function pointers and have the scheduler fill those in.

For the moment, the choice of scheduler is fixed to fair on the main CPU and staircase on secondary CPUs, haven't decided on a mechanism for changing that yet.

Updated all the users of the scheduling to reflect the changes.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>