Added the first version of the wait graph mechanism.
The wait graph mechanism is a new mechanism for tracking on what a particular yarn is waiting. When a yarn is suspended to wait on something, it is inserted into the wait graph along with one or more dependencies. When these dependencies are all satisfied, the yarn is resumed and rescheduled.

There are two types of dependencies: time and token. Time is obvious: after a specific time, it continues: this is handy for waits. Tokens are less obvious: tokens are numbers generated by the wait graph structure. Once a token is generated, dependencies can be inserted on that token, which will be fulfilled when a token is "signalled". This is handy for implementing things like locks, semaphores, condition locks, or other dependencies between yarns.

The rest of the system does not yet use this.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>